requirements.txt
google-adk==1.0.0
mcp==1.9.1
Deprecated ==1.2.13
python-dotenv==1.1.0